fix USD value not showing
